[Refactor] Remove jax suffix from file names
