+++ tpu_inference/runner/persistent_batch_manager.py
@@ -4,7 +4,7 @@ import jax
 from vllm.v1.core.sched.output import SchedulerOutput as VllmSchedulerOutput
 
 from tpu_inference.logger import init_logger
-from tpu_inference.runner.input_batch_jax import CachedRequestState, InputBatch
+from tpu_inference.runner.input_batch import CachedRequestState, InputBatch
 
 logger = init_logger(__name__)
