+++ tpu_inference/worker/tpu_worker.py
@@ -28,7 +28,7 @@ from tpu_inference.distributed.utils import (get_host_ip, get_kv_transfer_port,
 from tpu_inference.layers.jax.sharding import ShardingConfigManager
 from tpu_inference.logger import init_logger
 from tpu_inference.runner.kv_cache import get_rpa_page_size_bytes
-from tpu_inference.runner.tpu_jax_runner import TPUModelRunner
+from tpu_inference.runner.tpu_runner import TPUModelRunner
 
 logger = init_logger(__name__)
